refactor: HtmlButton module to core
